# Optimize CPU Monitor: 25% perf boosts and other changes