--- a/src/Xmobar/Plugins/Monitors/Common/Output.hs
+++ b/src/Xmobar/Plugins/Monitors/Common/Output.hs
@@ -43,6 +43,7 @@ module Xmobar.Plugins.Monitors.Common.Output ( IconPattern
                                              , pShowPercentBar
                                              , pShowVerticalBar
                                              , pShowIconPattern
+                                             , pShowPercentWithColors
                                              ) where
 
 import Data.Char