… bench/main.hs
… test/Xmobar/Plugins/Monitors/CpuSpec.hs
… xmobar.cabal