Summary of the changes:

* Make the CPU monitor more type safe. Added various new types instead of having `[String]` and `[Float]` everywhere.
* Add test cases for the CPU monitor
* Introduce new type named `PureConfig`. This is a pure version of `MConfig` and the initial work has been done in this MR which will hopefully in the future remove the need of the whole `MConfig` type.
* Avoid template parsing for the components which won't be rendered
* Avoid formatting output for the components which won't be rendered
* Perf improvements:

Before:
```
benchmarked Cpu Benchmarks/CPU normal args
time                 101.6 μs   (100.8 μs .. 102.3 μs)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 103.0 μs   (102.4 μs .. 104.5 μs)
std dev              2.883 μs   (1.465 μs .. 5.154 μs)
variance introduced by outliers: 11% (moderately inflated)
```
After:
```
time                 75.80 μs   (75.15 μs .. 76.36 μs)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 75.97 μs   (75.74 μs .. 76.41 μs)
std dev              1.057 μs   (666.6 ns .. 1.927 μs)
```