@jaor 2020-06-14
This looks very good indeed.  I am very happy to see tests too.  I have just a nit, that might be a bit annoying, but here it goes: i don't like the `PureConfig` name, because 'pure' just describes an implementation detail, which is neither here nor there (it feels like "typed Hungarian notation" if you take my meaning).  How about we call it simply `MonitorConfig`?