@psibi 2020-06-14
@jaor Thanks for the feedback! I will do the changes tomorrow. Free feel to let me know other changes too (including naming!). :-)