@jaor 2020-06-23
merged, thanks! (even travis-ci is working now!)